requirements.txt
ultralytics
opencv-python
numpy
pillow
tk